recipe tab and new recipe tab
fixed the formatting(styling? spacing) for the new recipe tab, and it has nigh full functionality. just waiting on the backend team to finish up so it can be fully integrated
as for the recipe tab, maybe it should be changed a bit? but other than that it should be done